+++ tehreer-android/src/main/java/com/mta/tehreer/opentype/PostScriptTable.java
@@ -60,8 +60,8 @@ public class PostScriptTable {
         this.table = new StructTable(typeface, pointer);
     }
 
-    public long version() {
-        return table.readUInt32(VERSION);
+    public int version() {
+        return table.readInt32(VERSION);
     }
 
     public int italicAngle() {
